+++ fessql/__init__.py
@@ -28,4 +28,4 @@ __all__ = (
     "__version__",
 )
 
-__version__ = "1.0.3b3"
+__version__ = "1.0.4"